chore: separate out utils from image input
